+++ src/StorageBase/Bridge/Contract/StorageFacadeInterface.php
@@ -10,6 +10,7 @@ use Heptacom\HeptaConnect\Storage\Base\Contract\Action\Identity\IdentityMapActio
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\Identity\IdentityOverviewActionInterface;
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\Identity\IdentityPersistActionInterface;
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\Identity\IdentityReflectActionInterface;
+use Heptacom\HeptaConnect\Storage\Base\Contract\Action\IdentityDirection\IdentityDirectionDeleteActionInterface;
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\IdentityError\IdentityErrorCreateActionInterface;
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\Job\JobCreateActionInterface;
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\Job\JobDeleteActionInterface;
@@ -68,6 +69,13 @@ interface StorageFacadeInterface
      * @throws StorageFacadeServiceExceptionInterface
      */
     public function getFileReferencePersistRequestAction(): FileReferencePersistRequestActionInterface;
+
+    /**
+     * Provides storage action to delete directional identities.
+     *
+     * @throws StorageFacadeServiceExceptionInterface
+     */
+    public function getIdentityDirectionDeleteAction(): IdentityDirectionDeleteActionInterface;
 
     /**
      * Provides storage action to map entities to their identities.